refactor: Member 패키지에서 Auth 패키지 분리
