+++ src/main/java/nextstep/auth/application/dto/GithubTokenRequest.java
@@ -1,4 +1,4 @@
-package nextstep.member.application.dto;
+package nextstep.auth.application.dto;
 
 public class GithubTokenRequest {
     private String code;
