+++ src/main/java/nextstep/auth/ui/TokenController.java
@@ -1,10 +1,10 @@
-package nextstep.member.ui;
+package nextstep.auth.ui;
 
-import nextstep.member.application.TokenService;
-import nextstep.member.application.dto.TokenDto;
-import nextstep.member.ui.dto.TokenRequestBody;
-import nextstep.member.ui.dto.TokenResponseBody;
-import nextstep.member.ui.dto.TokenFromGithubRequestBody;
+import nextstep.auth.application.TokenService;
+import nextstep.auth.application.dto.TokenDto;
+import nextstep.auth.ui.dto.TokenRequestBody;
+import nextstep.auth.ui.dto.TokenResponseBody;
+import nextstep.auth.ui.dto.TokenFromGithubRequestBody;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -18,14 +18,14 @@ public class TokenController {
         this.tokenService = tokenService;
     }
 
-    @PostMapping("/login/token")
+    @PostMapping("/auth/login")
     public ResponseEntity<TokenResponseBody> createToken(@RequestBody TokenRequestBody request) {
         TokenDto token = tokenService.createToken(request.getEmail(), request.getPassword());
 
         return ResponseEntity.ok(TokenResponseBody.create(token.getAccessToken()));
     }
 
-    @PostMapping("/login/github")
+    @PostMapping("/auth/login/github")
     public ResponseEntity<TokenResponseBody> createTokenFromGithub(@RequestBody TokenFromGithubRequestBody request) {
         TokenDto token = tokenService.createTokenFromGithub(request.getCode());
