+++ src/main/java/nextstep/auth/ui/dto/TokenFromGithubRequestBody.java
@@ -1,4 +1,4 @@
-package nextstep.member.ui.dto;
+package nextstep.auth.ui.dto;
 
 public class TokenFromGithubRequestBody {
     private String code;
