+++ src/main/java/nextstep/favorite/ui/FavoriteController.java
@@ -7,7 +7,7 @@ import nextstep.favorite.application.dto.FavoriteFindQuery;
 import nextstep.favorite.ui.dto.FavoriteRequest;
 import nextstep.favorite.ui.dto.FavoriteResponse;
 import nextstep.member.domain.LoginMember;
-import nextstep.member.ui.AuthenticationPrincipal;
+import nextstep.auth.ui.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
